+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.1] - 2025-11-05
 ### Fixed
 - Fixed player crash when clearing app from recents ([#298])
 - Fixed overlap between current track and system bars
@@ -86,7 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#269]: https://github.com/FossifyOrg/Music-Player/issues/269
 [#298]: https://github.com/FossifyOrg/Music-Player/issues/298
 
-[Unreleased]: https://github.com/FossifyOrg/Music-Player/compare/1.5.0...HEAD
+[Unreleased]: https://github.com/FossifyOrg/Music-Player/compare/1.5.1...HEAD
+[1.5.1]: https://github.com/FossifyOrg/Music-Player/compare/1.5.0...1.5.1
 [1.5.0]: https://github.com/FossifyOrg/Music-Player/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/FossifyOrg/Music-Player/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/FossifyOrg/Music-Player/compare/1.2.3...1.3.0